Reorganize command handlers into store-specific directories
