+++ docstore_manager/common/command/__init__.py
@@ -1,7 +1,5 @@
 """Command handlers for document store operations."""
 
 from .base import DocumentStoreCommand, CommandResponse
-from .qdrant import QdrantCommand
-from .solr import SolrCommand
 
-__all__ = ['DocumentStoreCommand', 'CommandResponse', 'QdrantCommand', 'SolrCommand'] 
+__all__ = ['DocumentStoreCommand', 'CommandResponse']
